Moved the Merkle Tree root generation logic from the block.rs to merkle.rs file as per requested on #96
